ksmb: avoid to send duplicate oplock break notifications
This patch fixes xfstests generic/011.

if ksmbd sends duplicate notifications for a file,
cifs could increment the reference count of the file
for each notification, and not decrement the count
because of the failure of queue_work.

So even if the file i closed, cifs never send a
SMB2_CLOSE request for the file.

Signed-off-by: Hyunchul Lee <hyc.lee@gmail.com>